MarkdownToPDF: Split sample and discover local font families
- Split the sample into focused source files for options, parsing,
  rendering, tagging, fonts, theme, tests, and the Markdown model.
  This keeps Program.cs centered on the APDFL sample entry flow.
- Move font-family discovery to APDFL runtime enumeration so
  --list-font-families reflects what is actually available on the
  local machine instead of a fixed catalog.
- Keep the existing rendering behavior intact while adding UTF-8
  decode validation, explicit disposal for APDFL-owned resources,
  and cleaner verbose font-name display.
